[hotfix] Move test purpose example to e2e test module in java.
